--- modules/add-remove-listener.js
@@ -1,19 +1,0 @@
-/* eslint-disable */
-import { displayEmpty } from './display-empty.js';
-import { library } from './books-class.js';
-
-// Function to add click event to remove button remove book from DOM
-export const addRemoveListener = (book) => {
-  document.getElementById(`remove-${book.id}`).addEventListener('click', (e) => {
-    e.preventDefault();
-    book.removeBook();
-    localStorage.setItem('library', JSON.stringify(library));
-    if (!library.length) {
-      displayEmpty();
-    }
-    const bookID = document.getElementById(`book-${book.id}`);
-    if (bookID.parentNode) {
-      bookID.parentNode.removeChild(bookID);
-    }
-  });
-};
