+++ modules/dom-elements.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 // Get all relevant elements from the DOM
 export const booksSection = document.getElementById('library');
 export const form = document.getElementById('form-id');
