+++ modules/random-id.js
@@ -1,2 +1,3 @@
-/* eslint-disable */
-export const uid = () => Date.now().toString(36) + Math.random().toString(36).substr(2);
+const uid = () => Date.now().toString(36) + Math.random().toString(36).substr(2);
+
+export default uid;
